python/requirements.txt
langchain
langchain-community
langchain-openai
langchain-ollama
langchain_chroma